myproject/requirements.txt
Django==2.0.13
djangorestframework==3.9.4
djangorestframework-simplejwt==4.1.4
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2019.3
